+++ src/test/java/org/folio/sidecar/service/routing/RoutingServiceTest.java
@@ -20,6 +20,7 @@ import io.vertx.ext.web.RoutingContext;
 import jakarta.enterprise.inject.Instance;
 import jakarta.ws.rs.NotFoundException;
 import java.util.List;
+import org.folio.sidecar.configuration.properties.ModuleProperties;
 import org.folio.sidecar.integration.am.ApplicationManagerService;
 import org.folio.sidecar.service.ModulePermissionsService;
 import org.folio.sidecar.support.TestConstants;
@@ -44,17 +45,20 @@ class RoutingServiceTest {
   @Mock private ModuleBootstrapListener listener1;
   @Mock private ModuleBootstrapListener listener2;
   @Mock private ModulePermissionsService modulePermissionsService;
+  @Mock private EgressBootstrapService egressBootstrapService;
+  @Mock private ModuleProperties moduleProperties;
 
   @BeforeEach
   void setUp() {
     when(instanceRequestHandler.get()).thenReturn(requestHandler);
     routingService = new RoutingService(appManagerService, instanceRequestHandler, List.of(listener1, listener2),
-      modulePermissionsService);
+      modulePermissionsService, egressBootstrapService, moduleProperties);
   }
 
   @AfterEach
   void tearDown() {
-    verifyNoMoreInteractions(appManagerService, requestHandler, listener1, listener2, modulePermissionsService);
+    verifyNoMoreInteractions(appManagerService, requestHandler, listener1, listener2, modulePermissionsService,
+      egressBootstrapService, moduleProperties);
   }
 
   @Test
@@ -128,5 +132,48 @@ class RoutingServiceTest {
   void updateModuleRoutes_negative_moduleNotFound() {
     routingService.updateModuleRoutes("unknown_module");
     verifyNoInteractions(appManagerService, listener1, listener2, router, route);
+  }
+
+  @Test
+  void init_positive_tenantScopedLoadsIngressOnly() {
+    routingService.tenantScoped = true;
+    var bootstrap = TestConstants.MODULE_BOOTSTRAP;
+    when(appManagerService.getIngressBootstrap()).thenReturn(succeededFuture(bootstrap));
+    when(router.route("/*")).thenReturn(route);
+
+    var listenersOrder = inOrder(listener1, listener2);
+
+    routingService.init(router);
+
+    listenersOrder.verify(listener1).onModuleBootstrap(bootstrap.getModule(), INIT);
+    listenersOrder.verify(listener1).onRequiredModulesBootstrap(bootstrap.getRequiredModules(), INIT);
+    listenersOrder.verify(listener2).onModuleBootstrap(bootstrap.getModule(), INIT);
+    listenersOrder.verify(listener2).onRequiredModulesBootstrap(bootstrap.getRequiredModules(), INIT);
+    verify(router).route("/*");
+    verify(route).handler(requestHandler);
+    verify(modulePermissionsService).putPermissions(anySet());
+  }
+
+  @Test
+  void onDiscovery_positive_tenantScopedSelfIsNoOp() {
+    routingService.tenantScoped = true;
+    when(moduleProperties.getId()).thenReturn(TestConstants.MODULE_ID);
+
+    routingService.onDiscovery(TestConstants.MODULE_ID);
+
+    verify(moduleProperties).getId();
+    verifyNoInteractions(appManagerService);
+  }
+
+  @Test
+  void onDiscovery_positive_tenantScopedOtherRefreshesTenants() {
+    routingService.tenantScoped = true;
+    when(moduleProperties.getId()).thenReturn(TestConstants.MODULE_ID);
+
+    routingService.onDiscovery("mod-bar-0.5.1");
+
+    verify(moduleProperties).getId();
+    verify(egressBootstrapService).refreshAllTenants();
+    verifyNoInteractions(appManagerService);
   }
 }